+++ gcs/CHANGES.md
@@ -2,7 +2,10 @@
 
 ## Next
 1. Added HNS optimizations for mkdir, getFileInfo, list, rename and delete.
-2. [Feature] Add support for cloud logging and prefix invocation ID to logs for correlation.
+1. [Feature] Add support for cloud logging and prefix invocation ID to logs for correlation.
+
+## 3.0.12 - 2025-09-26
+1. Enable move by default
 
 ## 3.0.11 - 2025-09-09
 1. PR #1484 - [Feature] Enabled write checksum by default